feat: migrate flaky runner to in runner file
