+++ src/driver.ts
@@ -23,6 +23,7 @@ import {
 	exportGeneratedPdf,
 	exportJSONReport,
 } from '@pkg/export.ts'
+import { flakyRunner } from '@pkg/runner.ts'
 
 const driver = async (
 	{ browser }: TDriverParams,
@@ -195,73 +196,13 @@ const driver = async (
 					}
 				})
 
-				const flakyCases: Array<TDataResult> = []
-				;((runs: number) => {
-					for (let i = 0; i < runs; i++) {
-						service.cases.forEach((c: TDrowserServiceCase) => {
-							const start = performance.now()
-
-							if (typeof c === 'object') {
-								const method =
-									(builder as unknown as Record<string, Function>)[c.method]
-
-								if (typeof method === 'function') {
-									const methodPromise = method.call(builder)
-									let actualValue: unknown = null
-
-									methodPromise.then((v: unknown) => {
-										const assertFunction = assert[c.operator] as TAssertFunction
-										actualValue = v
-										assertFunction(actualValue, c.except)
-
-										const end = performance.now()
-
-										flakyCases.push(
-											result({
-												id: nanoid(),
-												name: c.method,
-												actual: actualValue,
-												exceptation: c.except,
-												status: caseStatus.passed,
-												timestamp: new Date(),
-												duration: end - start,
-												month_of_test: month,
-												type: dataResultType.object,
-												browser,
-											}),
-										)
-									})
-										.catch(() => {
-											const end = performance.now()
-
-											flakyCases.push(
-												result({
-													id: nanoid(),
-													name: c.method,
-													actual: actualValue,
-													exceptation: c.except,
-													status: caseStatus.failed,
-													timestamp: new Date(),
-													duration: end - start,
-													month_of_test: month,
-													type: dataResultType.object,
-													browser,
-												}),
-											)
-										})
-
-									methodPromises.push(methodPromise)
-								} else {
-									console.error(
-										`Method ${c.method} not found on builder object.`,
-									)
-								}
-							}
-
-							if (typeof c === 'function') {}
-						})
-					}
-				})(5)
+				const flakyCases = flakyRunner({
+					builder,
+					service,
+					browser,
+					result,
+					methodPromises,
+				})
 
 				const exportGeneratedFiles = () => {
 					if (exportPdf) exportGeneratedPdf({ results: data.results })
